+++ Pandemic/model/countries/Netherlands.cs
@@ -6,7 +6,7 @@ namespace Pandemic.model.countries
 {
     class Netherlands : Country
     {
-        public Netherlands(string name, long population, int amountOfDiseases, ColorEnum color) : base(name, population, amountOfDiseases, color)
+        public Netherlands(string name, uint population, int amountOfDiseases, ColorEnum color) : base(name, population, amountOfDiseases, color)
         {
         }
     }
